kernel_cache: address review on the docs, the tests and the wording
Documentation and comments no longer mention the engine configuration.
The user-visible fact is that two different shapes can use one entry, so
the entry count does not track the number of shapes. How cuDNN keys an
entry is an implementation detail that the frontend must not promise.

revision() is now described as the way to find if the cache changed, and
a new serialization is given as an example of what to do with that
answer, and not as the only purpose. size() is called out as an
unreliable measure, because cuDNN can replace data that is already in
the cache.

Drop the advice about not keeping a counter value with a serialized
cache. The statement that the counter is local to the process and that
from_json() starts a new count already carries it.

Drop the method list from the thread-safety paragraph. It goes stale each
time a method joins the class, as it did here.

Merge the kernel cache tests into one file and name it kernel_cache.cpp,
so the two suites share the graph builders. The builders take a shape,
which defaults to the shape that the thread-safety tests already used.

Give the existing build_cudnn_graph() helper a build_plans parameter
instead of adding a second helper beside it.
