+++ test/cpp/CMakeLists.txt
@@ -29,8 +29,7 @@ add_executable(
     version.cpp
     tensor.cpp
     get_engine_and_knobs.cpp
-    kernel_cache_thread_safety.cpp
-    kernel_cache_revision.cpp
+    kernel_cache.cpp
 )
 
 if (MSVC)
